requirements_gemini.txt
# Python Requirements for Agentic Turing Machine (Google Gemini Version)

# Google Generative AI SDK for Gemini API
google-generativeai>=0.3.0

# Scientific computing
numpy>=1.24.0

# Data visualization
matplotlib>=3.7.0

# Machine learning utilities (for cosine similarity)
scikit-learn>=1.3.0

